Added better SRU and effort allocation for the EPD VOCT used in serializing

+++ epd-common/src/main/java/dk/dma/epd/common/prototype/model/voct/SearchPatternGenerator.java
@@ -51,7 +51,7 @@ public class SearchPatternGenerator {
     }
 
     public SearchPatternRoute generateSearchPattern(searchPattern pattern,
-            SARData data, NavSettings settings, int i) {
+            SARData data, NavSettings settings, long i) {
 
         this.settings = settings;
